backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
boto3==1.34.0
datasets==2.16.0
Werkzeug==3.0.1
